+++ Pod/Core/StepDefiner.swift
@@ -51,11 +51,25 @@ open class StepDefiner: NSObject {
          }
      
      - parameter expression: The expression to match against
-     - parameter f1: The step definition to be run, passing in the matches from the expression
+     - parameter f: The step definition to be run, passing in the matches from the expression
      
      */
-    open func step(_ expression: String, file: String = #file, line: Int = #line, f1: @escaping ([String])->()) {
-        self.test.addStep(expression, file: file, line: line, f1)
+    open func step<T: MatchedStringRepresentable>(_ expression: String, file: String = #file, line: Int = #line, f: @escaping ([T])->()) {
+        self.test.addStep(expression, file: file, line: line) { (matches: [String]) in
+
+            // Convert the matches to the correct type
+            var converted = [T]()
+            for match in matches {
+                guard let convert = T(fromMatch: match) else {
+                    XCTFail("Failed to convert \(match) to \(T.self) in \"\(expression)\"")
+                    return
+                }
+
+                converted.append(convert)
+            }
+
+            f(converted)
+        }
     }
     
     /**
@@ -68,44 +82,21 @@ open class StepDefiner: NSObject {
          }
      
      - parameter expression: The expression to match against
-     - parameter f1s: The step definition to be run, passing in the first capture group from the expression
+     - parameter f1: The step definition to be run, passing in the first capture group from the expression
      */
-    open func step(_ expression: String, file: String = #file, line: Int = #line, f1s: @escaping (String)->()) {
+    open func step<T: MatchedStringRepresentable>(_ expression: String, file: String = #file, line: Int = #line, f1: @escaping (T)->()) {
         self.test.addStep(expression, file: file, line: line) { (matches: [String]) in
             guard let match = matches.first else {
                 XCTFail("Expected single match not found in \"\(expression)\"")
                 return
             }
-            
-            f1s(match)
-        }
-    }
-    
-    /**
-     If you only want to match the first parameter, this will help make your code nicer
-     
-     Don't pass anything for file: or path: - these will be automagically filled out for you. Use it like this:
-     
-     step("Some (regular|irregular) expression") { (match: Int) in
-     ... some function ...
-     }
-     
-     - parameter expression: The expression to match against
-     - parameter f1s: The step definition to be run, passing in the first capture group from the expression
-     */
-    open func step(_ expression: String, file: String = #file, line: Int = #line, f1i: @escaping (Int)->()) {
-        self.test.addStep(expression, file: file, line: line) { (matches: [String]) in
-            guard let match = matches.first else {
-                XCTFail("Expected single match not found in \"\(expression)\"")
+
+            guard let integer = T(fromMatch: match) else {
+                XCTFail("Could not convert \"\(match)\" to \(T.self)")
                 return
             }
-            
-            guard let integer = Int(match) else {
-                XCTFail("Could not convert \"\(match)\" to an integer")
-                return
-            }
-            
-            f1i(integer)
+
+            f1(integer)
         }
     }
     
@@ -114,55 +105,35 @@ open class StepDefiner: NSObject {
      
      Don't pass anything for file: or path: - these will be automagically filled out for you. Use it like this:
      
-         step("Some (regular|irregular) expression with a second capture group here (.*)") { (match1: String, match2: String) in
+         step("Some (regular|irregular) expression with a second capture group here (.*)") { (match1: String, match2: Int) in
              ... some function ...
          }
      
      - parameter expression: The expression to match against
-     - parameter f2s: The step definition to be run, passing in the first two capture groups from the expression
+     - parameter f2: The step definition to be run, passing in the first two capture groups from the expression
      */
-    open func step(_ expression: String, file: String = #file, line: Int = #line, f2s: @escaping (String, String)->()) {
+    open func step<T: MatchedStringRepresentable, U: MatchedStringRepresentable>(_ expression: String, file: String = #file, line: Int = #line, f2: @escaping (T, U)->()) {
         self.test.addStep(expression, file: file, line: line) { (matches: [String]) in
             
             guard matches.count >= 2 else {
                 XCTFail("Expected at least 2 matches, found \(matches.count) instead, from \"\(expression)\"")
                 return
             }
-            
-            f2s(matches[0], matches[1])
-        }
-    }
-    
-    /**
-     If you only want to match the first parameters as integers, this will help make your code nicer
-     
-     Don't pass anything for file: or path: - these will be automagically filled out for you. Use it like this:
-     
-     step("Some (regular|irregular) expression with a second capture group here (.*)") { (match1: Int, match2: Int) in
-     ... some function ...
-     }
-     
-     - parameter expression: The expression to match against
-     - parameter f2i: The step definition to be run, passing in the first two capture groups from the expression
-     */
-    open func step(_ expression: String, file: String = #file, line: Int = #line, f2i: @escaping (Int, Int)->()) {
-        self.test.addStep(expression, file: file, line: line) { (matches: [String]) in
-            
-            guard matches.count >= 2 else {
-                XCTFail("Expected at least 2 matches, found \(matches.count) instead, from \"\(expression)\"")
+
+            guard let i1 = T(fromMatch: matches[0]) else {
+                XCTFail("Could not convert '\(matches[0])' to \(T.self), from \"\(expression)\"")
+                return
+            }
+
+            guard let i2 = U(fromMatch: matches[1]) else {
+                XCTFail("Could not convert '\(matches[1])' to \(U.self), from \"\(expression)\"")
                 return
             }
             
-            guard let i1 = Int(matches[0]),
-                let i2 = Int(matches[1]) else {
-                    XCTFail("Could not convert matches (\(matches[0]) and \(matches[1])) to integers, from \"\(expression)\"")
-                    return
-            }
-            
-            f2i(i1, i2)
+            f2(i1, i2)
         }
     }
-    
+
     /**
      Run other steps from inside your overridden defineSteps() method.