+++ CHANGELOG.md
@@ -1,8 +1,17 @@
 # XCTest-Gherkin changelog
 
 ### Unreleased
++ Clearer failure message when step isn't found
++ Allow Double and Bool as closure types in step definitions
++ Allow mix of closure parameter types in step definitions with two matches
+
+### 0.10.2
 + Update .travis.yml to Xcode 8.3
 + Update CocoaPods version in .travis.yml to 1.2.0
+
+### 0.10.1
+## 0.10
+Please prefer 0.10.2
 
 ## 0.9
 + Fix for Xcode 8.2